Refactor location of coursePage
